front-end files and folders organized
