+++ src/assets/Products/Card/Card-Elements/01-getItemCardName.js
@@ -1,0 +1,9 @@
+export const getItemCardName = ( name ) => {
+    
+    const itemName = document.createElement('div');
+    itemName.classList.add('product-name');
+    itemName.innerHTML = `${name}`;
+
+    return itemName;
+    
+}
