+++ src/assets/Products/Card/Card-Elements/02-getItemCardImage.js
@@ -1,0 +1,12 @@
+export const getItemCardImage = ( url_image ) => {
+    
+    const imgContainer = document.createElement('div');
+    imgContainer.classList.add('card-img');
+        
+        const img = document.createElement('img');
+        img.src = `${url_image}`;
+        
+    imgContainer.appendChild(img);
+
+    return imgContainer;
+}
