+++ src/index.html
@@ -10,12 +10,13 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-OERcA2EqjJCMA+/3y+gxIOqMEjwtxJY7qPCqsdltbNJuaOe923+mo//f6V8Qbsw3" crossorigin="anonymous"></script>
 
     <title>BSALE SHOP</title>
+    
 </head>
 <body style="overflow: auto;">
+    
     <div class="root">
         <div class="section-div"></div>
-    </div>
-    
+    </div>  
     
     <script type="module" src="index.js"></script> 
 </body>
